Fix death bug when no groups are set Fix reputation calculation after death Change reputation calculation method by ignoring HP loss Backward change for 'delete' button in adminActionFix Forgotten translations
Reputation calculation : HP loss could lead to a 0 reputation, but when
player died, he came back with full HP, hence a more than 0 reputation
(based on XP). That was weird.
Delete button : this is a faster way to delete actions, especially death
action (which is deleted in all team players' history pages)