Fix PHP memory crash because of fight requests
If a player requests a fight for an exercise and this player has the
best time, is the most trained and has many records, training page
became unloadable because of memory usage.
The workout consisted in new backend strategy : replacing mostTrained,
bestPlayer and fightRequest Page fields by bestTrainedPlayerId,
bestTimePlayerId and fight_request Text fields.

Seems like there is some kind of recursive issue between Page fields : a
Page reference in an 'exercise' template pointing to a 'player' template
page pointing in its turn to the 'exercise' template page through the
'fight request' field.